refactor: Member 관심사 분리
