+++ src/main/java/com/mjuAppSW/joA/domain/member/infrastructure/PasswordManager.java
@@ -1,0 +1,12 @@
+package com.mjuAppSW.joA.domain.member.infrastructure;
+
+public interface PasswordManager {
+
+    String createHashed(String rawPassword, String salt);
+
+    String createSalt();
+
+    String createRandom();
+
+    void validate(String rawPassword);
+}
